# Replace `compress/gzip` with `klauspost/compress`