--- a/go.mod
+++ b/go.mod
@@ -5,6 +5,7 @@ go 1.21
 require (
 	github.com/elastic/go-elasticsearch/v8 v8.12.1
 	github.com/json-iterator/go v1.1.12
+	github.com/klauspost/compress v1.17.7
 	github.com/stretchr/testify v1.9.0
 	go.elastic.co/apm/module/apmelasticsearch/v2 v2.5.0
 	go.elastic.co/apm/module/apmzap/v2 v2.5.0
@@ -27,7 +28,6 @@ require (
 	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/google/go-cmp v0.6.0 // indirect
 	github.com/joeshaw/multierror v0.0.0-20140124173710-69b34d4ec901 // indirect
-	github.com/klauspost/compress v1.17.7 // indirect
 	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
 	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/pkg/errors v0.9.1 // indirect